Move NetworkOff icon into Icons
