+++ libs/ui/src/icons.tsx
@@ -1,3 +1,4 @@
+import { useId } from 'react';
 import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import {
   faCircle,
@@ -632,6 +633,55 @@ export const Icons = {
 
   Network(props) {
     return <FaIcon {...props} flipInRtlMode={false} type={faSitemap} />;
+  },
+
+  NetworkOff(props: Pick<IconProps, 'className' | 'color' | 'style'>) {
+    const { className, color, style = {} } = props;
+    const theme = useTheme();
+    const maskId = useId();
+    return (
+      <StyledSvgIcon
+        aria-hidden="true"
+        className={[FONT_AWESOME_INLINE_SVG_CLASS_NAME, className]
+          .filter(Boolean)
+          .join(' ')}
+        data-icon="network-off"
+        role="img"
+        viewBox="0 0 576 512"
+        style={{
+          width: '1.125em',
+          color: iconColor(theme, color),
+          ...style,
+        }}
+      >
+        <mask
+          id={maskId}
+          maskUnits="userSpaceOnUse"
+          x="0"
+          y="-64"
+          width="576"
+          height="640"
+        >
+          <rect y="-64" width="576" height="640" fill="white" />
+          <path
+            d="M53 4L523 508"
+            stroke="black"
+            strokeWidth="110"
+            strokeLinecap="round"
+          />
+        </mask>
+        <path
+          mask={`url(#${maskId})`}
+          d="M208 80c0-26.5 21.5-48 48-48l64 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-8 0 0 40 152 0c30.9 0 56 25.1 56 56l0 32 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-32c0-4.4-3.6-8-8-8l-152 0 0 40 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-40-152 0c-4.4 0-8 3.6-8 8l0 32 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-32c0-30.9 25.1-56 56-56l152 0 0-40-8 0c-26.5 0-48-21.5-48-48l0-64z"
+        />
+        <path
+          d="M94 48L482 464"
+          stroke="currentColor"
+          strokeWidth="54"
+          strokeLinecap="round"
+        />
+      </StyledSvgIcon>
+    );
   },
 
   Sort(props) {
