+++ libs/ui/src/network_status_indicator.tsx
@@ -1,6 +1,5 @@
-import { useId } from 'react';
-import styled, { useTheme } from 'styled-components';
-import { iconColor, IconProps, Icons } from './icons';
+import styled from 'styled-components';
+import { Icons } from './icons';
 
 /**
  * The network statuses a machine's toolbar indicator can display, grouped by
@@ -37,60 +36,6 @@ const IndicatorRow = styled.div`
   white-space: nowrap;
 `;
 
-/**
- * A slashed variant of the network (sitemap) icon, used for warning and error
- * states. Not part of FontAwesome, so it's inlined here. Sized and aligned to
- * match the FontAwesome icons rendered by `Icons`.
- */
-const NetworkOffSvg = styled.svg`
-  height: 1em;
-  width: 1.125em;
-  vertical-align: -0.125em;
-`;
-
-function NetworkOffIcon({ color }: Pick<IconProps, 'color'>): JSX.Element {
-  const maskId = useId();
-  const theme = useTheme();
-  return (
-    <NetworkOffSvg
-      data-testid="network-off-icon"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 576 512"
-      fill="none"
-      aria-hidden="true"
-      style={{ color: iconColor(theme, color) }}
-    >
-      <mask
-        id={maskId}
-        maskUnits="userSpaceOnUse"
-        x="0"
-        y="-64"
-        width="576"
-        height="640"
-      >
-        <rect y="-64" width="576" height="640" fill="white" />
-        <path
-          d="M53 4L523 508"
-          stroke="black"
-          strokeWidth="110"
-          strokeLinecap="round"
-        />
-      </mask>
-      <path
-        mask={`url(#${maskId})`}
-        fill="currentColor"
-        d="M208 80c0-26.5 21.5-48 48-48l64 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-8 0 0 40 152 0c30.9 0 56 25.1 56 56l0 32 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-32c0-4.4-3.6-8-8-8l-152 0 0 40 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-40-152 0c-4.4 0-8 3.6-8 8l0 32 8 0c26.5 0 48 21.5 48 48l0 64c0 26.5-21.5 48-48 48l-64 0c-26.5 0-48-21.5-48-48l0-64c0-26.5 21.5-48 48-48l8 0 0-32c0-30.9 25.1-56 56-56l152 0 0-40-8 0c-26.5 0-48-21.5-48-48l0-64z"
-      />
-      <path
-        d="M94 48L482 464"
-        stroke="currentColor"
-        strokeWidth="54"
-        strokeLinecap="round"
-      />
-    </NetworkOffSvg>
-  );
-}
-
 export type NetworkStatusIndicatorProps =
   | { isHost: true; status: HostNetworkIndicatorStatus }
   | { isHost?: false; status: NetworkIndicatorStatus };
@@ -113,16 +58,15 @@ export function NetworkStatusIndicator(
       label: isHost ? 'Network Online' : 'Connected',
     },
     'no-host-connected': {
-      icon: <NetworkOffIcon color="inverse" />,
-
+      icon: <Icons.NetworkOff color="inverse" />,
       label: 'No VxAdmin Connected',
     },
     'no-network': {
-      icon: <NetworkOffIcon color="inverse" />,
+      icon: <Icons.NetworkOff color="inverse" />,
       label: 'No Network',
     },
     error: {
-      icon: <NetworkOffIcon color="inverseDanger" />,
+      icon: <Icons.NetworkOff color="inverseDanger" />,
       label: 'Network Error',
     },
   };
